update to version 2
