+++ package.json
@@ -6,19 +6,21 @@
   "dependencies": {
     "autoprefixer": "^8.6.2",
     "dotenv": "^5.0.1",
-    "gatsby": "^1.9.270",
+    "gatsby": "^2.0.106",
     "gatsby-cli": "^1.1.58",
-    "gatsby-link": "^1.6.44",
-    "gatsby-plugin-google-analytics": "^1.0.31",
-    "gatsby-plugin-postcss-sass": "^1.0.21",
-    "gatsby-plugin-react-helmet": "^2.0.11",
+    "gatsby-plugin-google-analytics": "^2.0.12",
+    "gatsby-plugin-react-helmet": "^3.0.5",
+    "gatsby-plugin-sass": "^2.0.8",
     "gatsby-plugin-styled-components": "^2.0.11",
     "gatsby-source-contentful": "^1.3.52",
     "gatsby-transformer-remark": "^1.7.42",
+    "node-sass": "^4.11.0",
     "postcss-clean": "^1.1.0",
     "postcss-custom-media": "^6.0.0",
     "postcss-custom-properties": "^7.0.0",
     "postcss-import": "^11.1.0",
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
     "react-headroom": "^2.2.2",
     "react-helmet": "^5.2.0",
     "react-redux": "^5.0.7",
@@ -40,7 +42,7 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "devDependencies": {
-    "babel-eslint": "^8.2.3",
+    "babel-eslint": "^8.2.6",
     "eslint": "^4.19.1",
     "eslint-config-prettier": "^2.9.0",
     "eslint-config-recommended": "^2.0.0",
