+++ src/components/card.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Link from 'gatsby-link'
+import { Link }  from 'gatsby'
 import styled from 'styled-components'
 import { media } from '../styles/style_utils'
