+++ src/components/footer.js
@@ -141,4 +141,4 @@ const Footer = () => (
   </div>
 )
 
-export default { Footer }
+export default  Footer
