+++ src/components/menu.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Link from 'gatsby-link'
+import { Link } from 'gatsby'
 import styled from 'styled-components'
 import Headroom from 'react-headroom'
 import {
@@ -141,4 +141,4 @@ const Menu = ({ menuIsOpen, toggleMenu, gradient }) => {
     </Splash>
   )
 }
-export default { BurgerWhite, BurgerGradient, Menu }
+export { BurgerWhite, BurgerGradient, Menu }
