+++ src/pages/our-work/cast.js
@@ -2,6 +2,7 @@ import React from 'react'
 import styled, { css } from 'styled-components'
 
 import Header from '../../containers/header-work'
+import Layout from '../../components/layout'
 import { SectionHeading, SectionTitle, Section } from '../../components/section'
 import {
   Wrapper,
@@ -38,7 +39,7 @@ const PictureWithFilter = styled(Picture)`
 `
 
 const AboutWork = () => (
-  <div>
+  <Layout>
     <Header
       topicLine="partnerships"
       tagLine="CAST Digital Fellowship"
@@ -170,7 +171,7 @@ const AboutWork = () => (
         </Wrapper>
       </div>
     </Section>
-  </div>
+  </Layout>
 )
 
 export default AboutWork
