+++ src/pages/our-work/coming-soon.js
@@ -2,6 +2,7 @@ import React from 'react'
 import styled, { css } from 'styled-components'
 
 import Header from '../../containers/header-work'
+import Layout from '../../components/layout'
 import { SectionHeading, SectionTitle, Section } from '../../components/section'
 import {
   Wrapper,
@@ -25,7 +26,7 @@ const PictureWithFilter = styled(Picture)`
 `
 
 const AboutWork = () => (
-  <div>
+  <Layout>
     <Header
       topicLine="Work in progress"
       tagLine="Coming soon!"
@@ -87,7 +88,7 @@ const AboutWork = () => (
         </Wrapper>
       </div>
     </Section>
-  </div>
+  </Layout>
 )
 
 export default AboutWork
