+++ src/state/state.js
@@ -1,5 +1,5 @@
 import { createStore as reduxCreateStore } from 'redux'
-import { menuReducer } from './reducers'
+import menuReducer from './reducers'
 
 const initialState = { menuIsOpen: 'INIT' }
